deleted use-device-type hook
